Animation class structure 008 camera rotation fixed

+++ ROD_main/Framework.cs
@@ -77,7 +77,7 @@ namespace ROD_engine_DX11
         public int frame_width;
         public int frame_height;
 
-        private long last_time;
+        private float last_time;
         
 
         public Framework(string title, int width, int height, bool depth, bool stencil, bool _VSync)
@@ -235,15 +235,13 @@ namespace ROD_engine_DX11
             // While there are no system messages, keep rendering
             while (IsIdle)
             {
-                long time = Stopwatch.ElapsedTicks / Stopwatch.Frequency;
-                long delta = time - last_time;
+                float time = Stopwatch.ElapsedTicks / (float)Stopwatch.Frequency;
+                float delta = time - last_time;
                 last_time = time;
-                /*
                 if (delta < 100)
                 {
                     Thread.Sleep(10);
                 }
-                 * */
                 MouseUpdate(time, delta);
                 KeyboardUpdate(time, delta);
                 Update(time, delta);
@@ -312,10 +310,10 @@ namespace ROD_engine_DX11
             keyboardState = new SharpDX.DirectInput.KeyboardState();            
         }
 
-        protected abstract void MouseUpdate(long time, long step);
-        protected abstract void KeyboardUpdate(long time, long step);
-        protected abstract void Update(long time, long step);
+        protected abstract void MouseUpdate(float time, float step);
+        protected abstract void KeyboardUpdate(float time, float step);
+        protected abstract void Update(float time, float step);
 
-        protected abstract void Render(long time, long step);
+        protected abstract void Render(float time, float step);
     }
 }
